requirements.txt
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
boto3==1.35.0
certifi==2025.8.3
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mypy==1.7.0
mypy_extensions==1.1.0
openai==1.3.0
playwright==1.52.0
passlib[bcrypt]==1.7.4
pydantic==2.11.9
pydantic_core==2.33.2
pyee==13.0.0
python-dotenv==1.0.0
sniffio==1.3.1
starlette==0.27.0
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.24.0
